Add a materialized view for doing search
This view combines information from all the fields that we want to be
able to perform a full text search on, with varying weights based on my
guess as to their importance.  The text is processed with the "english"
standard search configuration, and stored in a column (named "tsv").
We'll then be able to search over this column directly.

The view includes information joined from tag and tag synonym tables.

An index are created on the primary key in the view, to allow concurrent
update.  A separate GIN index is created on the tsv field to maintain
performance as the collection gets larger.

We also create a table, a pair of functions and some triggers to help maintain
the materialized view.  Firstly, the activities_search_is_dirty table is
created and simply contains a column.  We set up a set of triggers to
append a new row to this table whenever any change is made to any of the
tables involved in the query which generates the materialized view.  The
data added to the row is actually irrelevant; we just put a single
boolean column in the schema so that there is something to add.

Secondly, we create the update_activities_search() function, which
checks if there are any rows in the activities_search_is_dirty table,
and if so first clears the rows, and then triggers a concurrent update
of the materialized view.

We don't just trigger the update of the materialized view directly
because this would result in many concurrent attempts to update the view
happening when making multiple changes to the database (eg, when
importing data).  This results in errors being thrown due to memory
exhaustion.

Instead, we'll use a cron job to update the materialized view when
needed.  Updating the materialized view is fairly expensive, though
(because the full join query to generate it has to be run, even if no
changes have happened), which is why we use the
activities_search_is_dirty table to identify when changes have actually
been made and it's worth updating.