Use the materialized view for searching
After a few attempts to use django.contrib.postgres.search for this, I
couldn't get good results with the available features, so I used SQL
directly.

In particular; we want documents to be returned if they only match some
of the search words.  Postgres provides the plainto_tsquery() function
for parsing user queries, but this only supports "AND"ing all the words
together.  What I wanted to do was use a query like this, and combine
the scores returned with an "OR" query (working around some weakness in
postgreses ranking formula). The best I've come up with so far for an OR
query is to strip out any non alpha-numerics from the query string, then
convert spaces to "|" operators, then process it with the raw
"to_tsquery" function that postgres provides (which should be safe to
do, since the stripping of non alpha-numerics will remove any query
operators from the text being parsed).  This produces an OR query, but
it's still worth combining with the default query, since the default
query is parse in a more sophisticated way and may contain some things
which have been stripped out by the very naive OR query generation
process.

We just return a list of all the Activity objects matching the search.
This is not very efficient, but the database size is small enough right
now that it's okay.  Really, we need to use pagination information to
ensure we only fetch the part of the result set we need.